configs/zynq_pluto_defconfig: use buildroot for kernel build
This leverages buildroot's ability to build the kernel
using the same env as for rootfs.

All that's needed is to point to the ADI kernel git repo
and specify a hash/commit/branch, specify which
DTB files to be generated.

This removes the need for any inhouse scripts to build
the kernel when working with buildroot, since buildroot
may have some mechanism to protect against external
vars.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>