configs/zynq_pluto_defconfig: use buildroot's toolchain
This will add some additional build-time for downloading
the GCC toolchain from Linaro [which is the default for ARM].

But the added benefit is one less configuration step
for the user.
The aim is to not have to specify any CROSS_COMPILE
variable.

Also, this makes it less of a pain to maintain, since
buildroot seems to ignore some env-vars [like CROSS_COMPILE]
in certain build configurations.
For example, buildroot's Makefile makes does a
`unexport CROSS_COMPILE` along with ARCH and other vars
we set in the top `plutosdr-fw` Makefile.

That means we end up [and may continue to end up]
fighting/hacking buildroot.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>