configs/zynq_pluto_defconfig: use buildroot's external toolchain
This will add some additional build-time for downloading
the GCC toolchain from Linaro [which is the default for ARM].
But, it's one time and it's short compared to building
the toolchain from scratch (which is also possible with buildroot).

The added benefit is one less configuration step
for the user.
The aim is to not have to specify any CROSS_COMPILE
variable.

Also, this makes it less of a pain to maintain, since
buildroot seems to ignore some env-vars [like CROSS_COMPILE].
For example, buildroot's Makefile makes does a
`unexport CROSS_COMPILE` along with ARCH and other vars
we set in the top `plutosdr-fw` Makefile.

It also seems that buildroot started to do
validation of kernel headers and GCC version.
This means that a user must sync his
`BR2_TOOLCHAIN_EXTERNAL_GCC_X_Y` and
`BR2_TOOLCHAIN_EXTERNAL_HEADERS_X_Y=y`
with the Vivado version, and that makes
buildroot less nice to use.

Another option would have been to hack up our
buildroot but that would mean a continuous
fighting/hacking with buildroot.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>